implements scoring
